Remove #last_manifest
This commit is the first step in fixing some internal logic errors.
The code previously used two manifests. These were both the same
file even though the last manifest was only ever read and never written.
The current manifest was never read and only written. This caused
problems. This commit changes DynamicFileTask to use only one manifest.

It also changes DynamicFileTask to be more resilient to internal
mistakes. This commit assures that a manifest is assigned to a
DynamicFileTask instance. Previously empty manifests were created as
needed. This was a problem because the proper manifest was not being
assigned correctly. This commit also changes DynamicFileTask#needed? to
raise an error when there is no manifest. This ensures that this
implementation error will never happen. It does make tests more verbose,
but I think the change is worth it