Refactor code to reduce manifest passing
This commit introduces Filer#manifest which returns the set manifest or
the pipeline's manifest. This makes testing easier because you can use
Filer#manifest= to set on for that test. Pipeline#manifest delegates to
the project.

I think there is still problems with this code. The connection between
Project, Pipeline, Filter, and DynamicFileTask is too complicated. This is
especially annoying because DynamicFileTask requires a manifest to
function properly. However, the manifest must be passed down from the
project (then through a pipeline and into a filter) which builds the
actual DynamicFileTask instances. I think this is a code smell.
Introducing more collaboraters is making the tests more complex. Using a
singleton is one possible solution. I will think about this later.