Fix errors in gsub tests and gsub filter
@block was written incorrectly in the previous commit. This commit
also fixes test issues against that occur when running tests in random
order. Setting a fresh rake application each time fixes the issue. Tasks
are marked as already invoked when using the same application since the
same input and output files are used for every test.

This commit also fixes a problem with "global" regex variables in the
gsub block. $1, $2, $~ are supposed to be accessible inside the block.
This does not work when passing in a block from another context. $1 and
such are actually local variables not global variables. The matches are
passed in as block arguments instead. The documenation has been updated
to reflect this important quirk.